requirements.txt
flask>=2.0.1,<3.0.0
google-generativeai
python-dotenv>=0.19.1,<1.0.0